Bugfix: Halbtags-Absenz zog vollen statt halben Ferientag ab
Nutzer-Feedback: "nehme ich einen halben Tag Ferien zieht es einen
ganzen Tag ab, das sollte nicht sein". Employee.vacation_balance()
rechnete intern korrekt in 0.5-Schritten, aber EmployeeBalanceSerializer
deklarierte vacation_used_days/vacation_remaining_days als IntegerField
-- DRF schnitt beim Serialisieren der API-Antwort still auf int() ab
(24.5 -> 24), wodurch die im Frontend angezeigte Ferientage-Zahl einen
vollen statt einen halben Tag Abzug zeigte. Der bestehende Test prüfte
nur das Modell direkt und deckte diese Lücke nicht ab -- neuer
API-Level-Test schliesst sie.

+++ scheduling/serializers.py
@@ -227,8 +227,14 @@ class EmployeeBalanceSerializer(serializers.Serializer):
     is_provisional = serializers.BooleanField()
     vacation_year = serializers.IntegerField()
     vacation_entitlement_days = serializers.IntegerField()
-    vacation_used_days = serializers.IntegerField()
-    vacation_remaining_days = serializers.IntegerField()
+    # Bugfix (2026-08, Nutzer-Feedback: "halbtags Ferien zieht einen ganzen
+    # Tag ab"): used_days/remaining_days können seit den Halbtags-Absenzen
+    # 0.5-Schritte enthalten (Employee.vacation_balance()) -- als
+    # IntegerField wurde das beim Serialisieren stillschweigend zu int()
+    # abgeschnitten (24.5 -> 24), wodurch die API einen vollen statt einen
+    # halben Tag Abzug meldete, obwohl das Modell selbst korrekt rechnete.
+    vacation_used_days = serializers.FloatField()
+    vacation_remaining_days = serializers.FloatField()
 
 
 class WeeklyOvertimeSerializer(serializers.Serializer):
